Combined two preview files
