+++ .storybook/preview.tsx
@@ -1,6 +1,23 @@
 import React from 'react';
 import type { Preview, Decorator } from '@storybook/react';
-import preview from './preview';
+import './style.css';
+
+// Storybook parameters
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+    options: {
+      storySort: {
+        order: ['Intro', 'Layers', 'Controls', 'Interactions'],
+      },
+    },
+  },
+};
 
 // Wrap all stories in React.StrictMode
 export const decorators: Decorator[] = [
@@ -11,5 +28,5 @@ export const decorators: Decorator[] = [
   ),
 ];
 
-export * from './preview';
+// Provides Storybook with your global config.
 export default preview;
